+++ CHANGES.md
@@ -5,6 +5,9 @@ version 1.6.1 - soon come
 ----------------------
 See milestone [1.6.1](https://github.com/btrplace/scheduler/milestones/1.6.1)
 
+- More efficient event injection (see Issue #145)
+
+
 version 1.6.0 - 08 Feb 2017
 ----------------------
 See milestone [1.6.0](https://github.com/btrplace/scheduler/milestones/1.6.0)